refactor: 회원가입 로직을 Auth 패키지로 분리(#2)
