--- src/main/java/io/wisoft/avocadobackendhexagonal/domain/member/application/port/in/SignupUseCase.java
@@ -1,8 +1,0 @@
-package io.wisoft.avocadobackendhexagonal.domain.member.application.port.in;
-
-import io.wisoft.avocadobackendhexagonal.domain.member.application.port.in.command.SignupCommand;
-
-public interface SignupUseCase {
-
-    Long signup(final SignupCommand request);
-}
